backend/pyproject.toml
[project]
name = "seasussed-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "google-adk>=1.17",
  "google-genai>=1.0",
  "duckdb>=1.1",
  "pydantic>=2.9",
  "numpy>=2.4.2",
  "pandas>=3.0.1",
]

[dependency-groups]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "httpx>=0.27",
  "ruff>=0.8",
  "mypy>=1.13",
]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
